+++ AutoSuit/MyPluginInfo.cs
@@ -5,9 +5,9 @@ namespace AutoSuit
     {
         public const string PLUGIN_GUID = $"{PLUGIN_ORIGINAL_AUTHOR}.{PLUGIN_NAME}";
         public const string PLUGIN_NAME = "AutoSuit";
-        public const string USERS_PLUGIN_NAME = "Auto Suit";
+        public const string USERS_PLUGIN_NAME = "AutoSuit";
         public const string PLUGIN_VERSION = "0.0.6";
-        public const string PLUGIN_DESCRIPTION = "Automatically equips (or unequips) the suit when pressing the pressurize lever in the airlock. Clientside.";
+        public const string PLUGIN_DESCRIPTION = "Automatically equips (or unequips) the suit when pressing the pressurize lever in the airlock. Client-side, Mod-session required to enable.";
         public const string PLUGIN_ORIGINAL_AUTHOR = "18107";
         public const string PLUGIN_AUTHORS = "18107";
         public const string PLUGIN_THUNDERSTORE_ID = "";